Revert "Delete blync-light-manager-spec"
This reverts commit a8b4a2ed7f9173ca59add833d6f6a3f685c624d5.

I removed this file initially because I was trying to see if I could
build on CircleCi and Travis without it.

It's problematic because it involves 'node-hid', and code associated
with node-hid doesn't run on the ubuntu containers on these CI systems.

I have a plan for running these specs now, and so I'm bringing it back.